Move chronogram sources
From com.hepia.logisim to cbruch.logisim.gui.chronogram
