+++ src/main/java/com/cburch/logisim/gui/chronogram/chronodata/TimelineParam.java
@@ -27,15 +27,43 @@
  *       Yverdon-les-Bains, Switzerland
  *       http://reds.heig-vd.ch
  *******************************************************************************/
-package com.hepia.logisim.chronogui;
+package com.cburch.logisim.gui.chronogram.chronodata;
 
-import javax.swing.JPanel;
+public class TimelineParam {
 
-/**
- * LeftPanel and RightPanel extends ChronoPanelTemplate
- */
-public class ChronoPanelTemplate extends JPanel {
+	public static String[] units = { "Hz", "KHz", "MHz", "GHz" };
 
-	private static final long serialVersionUID = 1L;
-	// not used anymore for now
+	private String selectedUnit;
+	private String signalClkName;
+	private int frequency;
+
+	public TimelineParam(String timelineString) {
+		String[] elems = timelineString.split(":");
+		this.selectedUnit = elems[3];
+		this.signalClkName = elems[1];
+		this.frequency = Integer.parseInt(elems[2]);
+	}
+
+	public TimelineParam(String selectedUnit, String signalClkName,
+			int frequency) {
+		this.selectedUnit = selectedUnit;
+		this.signalClkName = signalClkName;
+		this.frequency = frequency;
+	}
+
+	public int getFrequency() {
+		return frequency;
+	}
+
+	public String getSelectedUnit() {
+		return selectedUnit;
+	}
+
+	public String getSignalClkName() {
+		return signalClkName;
+	}
+
+	public String toString() {
+		return "clk:" + signalClkName + ":" + frequency + ":" + selectedUnit;
+	}
 }
