+++ src/main/java/com/cburch/logisim/gui/chronogram/chronogui/LeftPanel.java
@@ -27,7 +27,7 @@
  *       Yverdon-les-Bains, Switzerland
  *       http://reds.heig-vd.ch
  *******************************************************************************/
-package com.hepia.logisim.chronogui;
+package com.cburch.logisim.gui.chronogram.chronogui;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
@@ -46,7 +46,7 @@ import javax.swing.SwingUtilities;
 import javax.swing.table.DefaultTableModel;
 import javax.swing.table.JTableHeader;
 
-import com.hepia.logisim.chronodata.SignalData;
+import com.cburch.logisim.gui.chronogram.chronodata.SignalData;
 
 /**
  * Chronogram's left side Panel Composed of one signalNameHead and multiple
