+++ src/main/java/com/cburch/logisim/gui/chronogram/chronogui/RightPanel.java
@@ -27,7 +27,7 @@
  *       Yverdon-les-Bains, Switzerland
  *       http://reds.heig-vd.ch
  *******************************************************************************/
-package com.hepia.logisim.chronogui;
+package com.cburch.logisim.gui.chronogram.chronogui;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
@@ -38,7 +38,7 @@ import javax.swing.Box;
 import javax.swing.JLayeredPane;
 import javax.swing.SwingUtilities;
 
-import com.hepia.logisim.chronodata.SignalData;
+import com.cburch.logisim.gui.chronogram.chronodata.SignalData;
 
 /**
  * Chronogram's right side Panel Composed of one TimeLine on top and multiple
